tests/distance_test/requirements.txt
opencv-python>=4.5.0
numpy>=1.20.0
mss>=6.0.0
pywin32>=300
